Implement HTML generation tasks
