+++ README.md
@@ -34,6 +34,11 @@ task exampleTask2(type: MarkdownToPdfTask){
 	inputFile = '/PATH/TO/CHANGELOG.md'
 	outputFile = '/PATH/TO/CHANGELOG.pdf'
 }
+
+task exampleTask3(type: MarkdownToHtmlTask){
+	inputFile = '/PATH/TO/CHANGELOG.md'
+	outputFile = '/PATH/TO/CHANGELOG.html'
+}
 ```
 Default Task Example:
 ```
@@ -41,11 +46,19 @@ Default Task Example:
 	-- build.gradle
 	-- README.md
 ```
-Here you can use the task "readmeToPdf" which will create
+Here the task "readmeToPdf" will create
 ```
 - example
 	-- build.gradle
 	-- README.md
 	-- build
 		---README.pdf
+```
+Here the task "readmeToHtml" will create
+```
+- example
+	-- build.gradle
+	-- README.md
+	-- build
+		---README.html
 ```
