+++ src/main/groovy/de/fntsoftware/gradle/MarkdownToHtmlTask.groovy
@@ -1,0 +1,11 @@
+package de.fntsoftware.gradle
+
+import org.gradle.api.tasks.TaskAction
+
+class MarkdownToHtmlTask extends AbstractMarkdownTask {
+	@TaskAction
+	void action() {
+		File file = new File(this.outputFile.path)
+		file << buildHtml()
+	}
+}
