+++ src/main/groovy/de/fntsoftware/gradle/MarkdownToPdfTask.groovy
@@ -1,62 +1,12 @@
 package de.fntsoftware.gradle
 
-import org.gradle.api.DefaultTask
-import org.gradle.api.InvalidUserDataException
-import org.gradle.api.tasks.InputFile
-import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.TaskAction
 
-import com.vladsch.flexmark.html.HtmlRenderer;
-import com.vladsch.flexmark.parser.Parser;
 import com.vladsch.flexmark.pdf.converter.PdfConverterExtension
-import com.vladsch.flexmark.util.options.MutableDataSet;
 
-class MarkdownToPdfTask extends DefaultTask {
-
-	/** Value which can be resolved using {@link org.gradle.api.Project#file(Object object)} */
-	private File inputFile
-
-	/** Value which can be resolved using {@link org.gradle.api.Project#file(Object object)} */
-	private File outputFile
-
-	@InputFile
-	File getInputFile() {
-		return this.inputFile
-	}
-
-	File setInputFile(Object file) {
-		this.inputFile = getProject().file(file)
-	}
-
-	@OutputFile
-	File getOutputFile() {
-		return this.outputFile
-	}
-
-	File setOutputFile(Object file) {
-		this.outputFile = getProject().file(file)
-	}
-
+class MarkdownToPdfTask extends AbstractMarkdownTask {
 	@TaskAction
 	void action() {
-		MarkdownToPdf settings = this.getProject().getExtensions().getByType(MarkdownToPdf.class);
-		if(this.inputFile.exists()) {
-			def outputDir = new File(outputFile.parent)
-			outputDir.mkdirs()
-
-			def options = new MutableDataSet()
-
-			def parser = Parser.builder(options).build()
-			def renderer = HtmlRenderer.builder(options).build()
-
-			def document = parser.parse(this.inputFile.getText())
-			def css = settings.getCssFileContent()
-			def html = "<html><head><style>${css}</style></head><body>${renderer.render(document)}</body></html>"
-
-			PdfConverterExtension.exportToPdf(this.outputFile.getPath(), html, "", options)
-		}
-		else {
-			throw new InvalidUserDataException("Inputfile does not exist.")
-		}
+		PdfConverterExtension.exportToPdf(this.outputFile.path, buildHtml(), "", options)
 	}
 }
